python-service/requirements.txt
# Core ML
torch>=2.2.0
torchvision>=0.17.0

# Unified CLIP-compatible embedding API
# - clip-ViT-B-32                                   image encoder (512-dim CLIP space)
# - clip-ViT-B-32-multilingual-v1  multilingual text encoder (same 512-dim CLIP space)
#   Supports EN, Japanese (日本語), Vietnamese (Tiếng Việt) and 50+ other languages.
sentence-transformers>=3.0.0

# FAISS vector similarity search
faiss-cpu>=1.8.0        # swap to faiss-gpu if CUDA is available

# Image processing
Pillow>=10.4.0

# Utilities
numpy>=1.26.0

# OCR text extraction — supports English, Japanese, Vietnamese and 80+ languages
easyocr>=1.7.0
